pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-news-linkedin-agent"
version = "0.1.0"
description = "AI agent that searches for latest AI news and posts to LinkedIn"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "APScheduler>=3.10.4",
    "linkedin-api>=2.0.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.setuptools]
packages = ["src"]
